backend/requirements.txt
Flask==2.0.2
pymongo==3.12.0
python-dotenv==0.19.1
